split cover serving code out to individual plugins
