+++ src/com/pugh/sockso/web/action/covers/CachedCoverer.java
@@ -1,10 +1,37 @@
 
 package com.pugh.sockso.web.action.covers;
 
-public class CachedCoverer implements CovererPlugin {
+import com.pugh.sockso.music.CoverArtCache;
+import java.io.IOException;
+
+public class CachedCoverer extends BaseCoverer {
     
-    public boolean serveCover( final String itemName ) {
+    /**
+     *  Try and serve a cached cover if there is one
+     * 
+     *  @param itemName
+     * 
+     *  @return
+     * 
+     *  @throws IOException 
+     * 
+     */
+
+    public boolean serveCover( final String itemName ) throws IOException {
+
+        final CoverArtCache coverCache = new CoverArtCache();
+
+        if ( coverCache.isCached(itemName) ) {
+            serveCover( 
+                coverCache.getCoverArt( itemName ),
+                itemName,
+                false
+            );
+            return true;
+        }
+
         return false;
+
     }
 
 }
